builder:  - Update WeightIndicatorResponse's weight from float to BigDecimal in builder

+++ builder/src/main/java/eu/esens/espdvcd/builder/model/ModelExtractor.java
@@ -577,7 +577,7 @@ public interface ModelExtractor {
             if (criterionType.getWeightNumeric() != null
                     && criterionType.getWeightNumeric().getValue() != null) {
 
-                response.setWeight(criterionType.getWeightNumeric().getValue().floatValue());
+                response.setWeight(criterionType.getWeightNumeric().getValue());
             }
         }
     }
